fix: 사용자 리뷰 작성 시 role 추가 #S10P12D208-230 (#in-progress)

+++ BackEnd/src/main/java/com/qzp/bid/domain/member/dto/MemberReviewReq.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class MemberReviewReq {
 
-    private int dealId; //거래 id
+    private long dealId; //거래 id
     private String targetNickname; //리뷰 대상자 닉네임
     private int score; //점수
     private String content; //내용
